+++ compiler/vec.h
@@ -61,7 +61,7 @@ template <class C> class Vec : public gc {
 };
 
 #define forv_Vec(_c, _p, _v) if ((_v).n) for (_c *qq__##_p = (_c*)0, *_p = (_v).v[0]; \
-                    ((int)(qq__##_p) < (_v).n) && ((_p = (_v).v[(int)qq__##_p]) || 1); qq__##_p = (_c*)(((int)qq__##_p) + 1))
+                    ((intptr_t)(qq__##_p) < (_v).n) && ((_p = (_v).v[(intptr_t)qq__##_p]) || 1); qq__##_p = (_c*)(((intptr_t)qq__##_p) + 1))
 
 // Intervals store sets in interval format (e.g. [1..10][12..12]).
 // Inclusion test is by binary search on intervals.
